Address review: own GitHub webhook replay protection in the GitHub agent
The delivery-dedup store lived in slack/state.ts and was only restored by
SlackAgent.startup, so a GitHub-only install (Slack disabled) served
/github/webhook with an empty in-memory map after a restart and re-ran
automations on GitHub's redeliveries. Move the store to github/deliveries.ts
and load it from GithubAgent.startup, before the webhook server can hand this
agent a delivery, so replay protection survives a restart without Slack.
