+++ packages/core/opensession-server/src/agents/slack/state.test.ts
@@ -14,9 +14,6 @@ const {
 	loadSession,
 	saveSession,
 	getSessionKey,
-	loadGithubDeliveries,
-	isGithubDeliveryProcessed,
-	markGithubDeliveryProcessed,
 } = await import("./state");
 const { writeJsonAtomic } = await import("../../server/shared/atomic-write");
 
@@ -78,20 +75,3 @@ describe("saveSession", () => {
 	});
 });
 
-describe("GitHub delivery replay protection", () => {
-	test("persists delivery ids and restores them after a reload", () => {
-		const deliveryId = "github-delivery-persists";
-		markGithubDeliveryProcessed(deliveryId);
-		expect(isGithubDeliveryProcessed(deliveryId)).toBe(true);
-
-		// loadGithubDeliveries clears the in-memory map first, mirroring a restart.
-		loadGithubDeliveries();
-		expect(isGithubDeliveryProcessed(deliveryId)).toBe(true);
-	});
-
-	test("drops expired delivery ids when restoring the persistent store", () => {
-		writeJsonAtomic(`${SESSION_DIR}/github-deliveries.json`, [["expired-delivery", 0]], false);
-		loadGithubDeliveries();
-		expect(isGithubDeliveryProcessed("expired-delivery")).toBe(false);
-	});
-});
